+++ springboot-samples/db/mybatis/base/src/main/resources/application.properties
@@ -1,7 +1,6 @@
 spring.application.name=base-mybatis
 logging.file.path=./db/mybatis/logs/
 logging.path=./db/mybatis/logs/
-logging.level.com.alipay.sofa=DEBUG
 logging.level.root=INFO
 logging.level.com.alipay.sofa.arklet=INFO
 logging.config=classpath:log4j2-spring.xml
