+++ springboot-samples/web/webflux/base/src/main/java/com/alipay/sofa/web/webflux/CountryHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.yuan.bizwebflux;
+package com.alipay.sofa.web.webflux;
 
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
@@ -25,14 +25,13 @@ import reactor.core.publisher.Mono;
 
 /**
  * @author: yuanyuan
- * @date: 2023/11/6 7:53 下午
+ * @date: 2023/11/6 5:28 下午
  */
 @Component
-public class BizHandler {
+public class CountryHandler {
 
-    public Mono<ServerResponse> helloBiz(ServerRequest request) {
+    public Mono<ServerResponse> helloCountry(ServerRequest request) {
         return ServerResponse.ok().contentType(MediaType.TEXT_PLAIN)
-            .body(BodyInserters.fromValue("Hello, biz webflux!"));
+            .body(BodyInserters.fromValue("Hello, country!"));
     }
-
 }
