+++ springboot-samples/web/webflux/base/src/test/java/com/alipay/sofa/web/webflux/DemowebfluxApplicationTests.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.yuan.bizwebflux;
+package com.alipay.sofa.web.webflux;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BizwebfluxApplicationTests {
+class DemowebfluxApplicationTests {
 
     @Test
     void contextLoads() {
