+++ springboot-samples/web/webflux/biz/src/main/java/com/alipay/sofa/web/webflux/BizHandler.java
@@ -14,16 +14,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.yuan.bizwebflux;
+package com.alipay.sofa.web.webflux;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.server.ServerRequest;
+import org.springframework.web.reactive.function.server.ServerResponse;
+import reactor.core.publisher.Mono;
 
-@SpringBootApplication
-public class BizwebfluxApplication {
+/**
+ * @author: yuanyuan
+ * @date: 2023/11/6 7:53 下午
+ */
+@Component
+public class BizHandler {
 
-    public static void main(String[] args) {
-        SpringApplication.run(BizwebfluxApplication.class, args);
+    public Mono<ServerResponse> helloBiz(ServerRequest request) {
+        return ServerResponse.ok().contentType(MediaType.TEXT_PLAIN)
+            .body(BodyInserters.fromValue("Hello, biz webflux!"));
     }
 
 }
