+++ springboot-samples/web/webflux/biz/src/main/java/com/alipay/sofa/web/webflux/RouterConfig.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.yuan.bizwebflux;
+package com.alipay.sofa.web.webflux;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
